+++ src/geometry/line_buffer.cpp
@@ -14,10 +14,10 @@ size_t LineVertexBuffer::add(vertex_type x, vertex_type y, float ex, float ey, i
     coords[1] = (y * 2) | ty;
 
     int8_t *extrude = static_cast<int8_t *>(data);
-    extrude[4] = std::round(extrudeScale * ex);
-    extrude[5] = std::round(extrudeScale * ey);
+    extrude[4] = static_cast<int8_t>(std::round(extrudeScale * ex));
+    extrude[5] = static_cast<int8_t>(std::round(extrudeScale * ey));
 
-    coords[3] = linesofar;
+    coords[3] = static_cast<int16_t>(linesofar);
 
     return idx;
 }
