+++ src/geometry/text_buffer.cpp
@@ -15,21 +15,21 @@ size_t TextVertexBuffer::add(int16_t x, int16_t y, float ox, float oy, uint16_t 
     int16_t *shorts = static_cast<int16_t *>(data);
     shorts[0] = x;
     shorts[1] = y;
-    shorts[2] = std::round(ox * 64); // use 1/64 pixels for placement
-    shorts[3] = std::round(oy * 64);
+    shorts[2] = static_cast<int16_t>(std::round(ox * 64)); // use 1/64 pixels for placement
+    shorts[3] = static_cast<int16_t>(std::round(oy * 64));
 
     uint8_t *ubytes = static_cast<uint8_t *>(data);
     // a_data1
-    ubytes[8] = tx / 4;
-    ubytes[9] = ty / 4;
-    ubytes[10] = labelminzoom * 10;
-    ubytes[11] = (int16_t)std::round(angle * angleFactor) % 256;
+    ubytes[8] = static_cast<uint8_t>(tx / 4);
+    ubytes[9] = static_cast<uint8_t>(ty / 4);
+    ubytes[10] = static_cast<uint8_t>(labelminzoom * 10);
+    ubytes[11] = static_cast<uint8_t>(std::round(angle * angleFactor)) % 256;
 
     // a_data2
-    ubytes[12] = minzoom * 10; // 1/10 zoom levels: z16 == 160.
-    ubytes[13] = std::fmin(maxzoom, 25) * 10; // 1/10 zoom levels: z16 == 160.
-    ubytes[14] = util::max((int16_t)std::round(range[0] * angleFactor), (int16_t)0) % 256;
-    ubytes[15] = util::min((int16_t)std::round(range[1] * angleFactor), (int16_t)255) % 256;
+    ubytes[12] = static_cast<uint8_t>(minzoom * 10); // 1/10 zoom levels: z16 == 160.
+    ubytes[13] = static_cast<uint8_t>(std::fmin(maxzoom, 25) * 10); // 1/10 zoom levels: z16 == 160.
+    ubytes[14] = static_cast<uint8_t>(util::max((int16_t)std::round(range[0] * angleFactor), (int16_t)0)) % 256;
+    ubytes[15] = static_cast<uint8_t>(util::min((int16_t)std::round(range[1] * angleFactor), (int16_t)255)) % 256;
 
     return idx;
 }
